convo/requirements.txt
pipecat-ai[cartesia,openai,silero,deepgram,elevenlabs]
fastapi
uvicorn
python-dotenv
loguru
twilio
yaml
